refactor(mllp-client): merge send+response → message.ts; fold reconnect → state.ts
Two file merges toward a simpler layout (12 → 10 files), no behavior change:
- send.ts + response.ts → message.ts — the HL7v2 codec (message ⇄ wire) in one
  module: toTree/toWireBytes/requestControlId (outbound) + parseResponse +
  MllpClientResponse (inbound).
- reconnect.ts → state.ts — the ReconnectPolicy/presets/backoffDelay live with
  the machine that consumes them.

99 tests pass.
