+++ packages/mllp-client/src/connection.ts
@@ -29,8 +29,8 @@ import { createFrameDecoder } from "@glion/mllp-transport";
 
 import type { MllpDuplex } from "./duplex";
 import { MllpClientError, MllpErrorCode, sendTimeoutError } from "./errors";
-import { parseResponse } from "./response";
-import type { MllpClientResponse } from "./response";
+import { parseResponse } from "./message";
+import type { MllpClientResponse } from "./message";
 
 /**
  * Maximum unsolicited frames buffered between sends. A well-behaved peer sends
